Use 170bp as a better estimate for wastewater than 120bp
I was trying to debug why I was seeing surprising results with the simulator, and I noticed that we've been using 120bp as an estimate for how short fragments tend to be when sequencing wastewater.  This is too conservative: looking at SARS-CoV-2 in the MU data I see a mean per-base coverage of 5.83 per 1000 reads (when excluding the amplicon spike contamination) over a 29,225bp genome, and 5.83 / 1000 * 29225 = 170.