Import wastewater distributions from p2ra
In https://www.medrxiv.org/content/10.1101/2023.12.22.23300450v3 and https://naobservatory.org/blog/predicting-influenza-abundance we generated distributions for RAi(1%) for SARS-CoV-2, influenza A, and influenza B.  In https://github.com/naobservatory/p2ra-restricted/pull/8 we exported them as a js file that the simulator could work with.  Make that js file available to the simulator.

Actually updating the simulator to use this data will be in a future PR